+++ rector.php
@@ -78,6 +78,8 @@ return RectorConfig::configure()
         __DIR__ . '/tests/Integration/Services/CRM/Documentgenerator/Template',
         __DIR__ . '/src/Services/Landing',
         __DIR__ . '/tests/Integration/Services/Landing',
+        __DIR__ . '/src/Services/Sign',
+        __DIR__ . '/tests/Integration/Services/Sign',
         __DIR__ . '/tests/Unit/',
     ])
     ->withCache(cacheDirectory: __DIR__ . '/var/.cache/rector')